refactor(data): split descriptor validators into own files; map Schema.ToType blob → Blob
Move IsValidWith{Partial,Complete}Descriptor out of is-valid.ts so the base
validator stays descriptor-agnostic. Fix Schema.ToType to resolve { type: "blob" }
to Blob instead of falling through to any; add a Blob-as-state-and-return test.
