+++ packages/data-lit-todo/package.json
@@ -1,6 +1,6 @@
 {
   "name": "data-lit-todo",
-  "version": "0.10.7",
+  "version": "0.10.6",
   "description": "Todo application - Lit web components with @adobe/data ECS",
   "type": "module",
   "private": true,
