+++ packages/data-react-hello/package.json
@@ -1,6 +1,6 @@
 {
   "name": "data-react-hello",
-  "version": "0.10.7",
+  "version": "0.10.6",
   "description": "Hello World sample - click counter using @adobe/data-react",
   "type": "module",
   "private": true,
