+++ packages/data/src/service/async-data-service/public.ts
@@ -1,4 +1,6 @@
 // © 2026 Adobe. MIT License. See /LICENSE for details.
 
 export * from "./is-valid.js";
+export * from "./is-valid-with-partial-descriptor.js";
+export * from "./is-valid-with-complete-descriptor.js";
 export * from "./create-lazy.js";
